backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
openai==1.57.2
sentence-transformers==3.3.1
pypdf==5.1.0
python-docx==1.1.2
numpy==2.1.3
rank-bm25==0.2.2
email-validator==2.2.0
aiosmtplib==3.0.1
pydantic[email]==2.10.3
